+++ lib/plexostables.h
@@ -387,7 +387,7 @@ union plexosKeyRef {
 
 struct plexosKeyIndex {
     int periodtype; // period_type_id
-    int position; // bytes from binary file start
+    long int position; // bytes from binary file start
     int length; // in 8-byte (64-bit float) increments
     int periodoffset; // period_offset // temporal data offset (if any) in stored times
     union plexosKeyRef key; // key_id
